chore: extract api in separate hook
