+++ package/src/components/Notifications/hooks/index.ts
@@ -1,4 +1,5 @@
 export * from './useNotificationApi';
+export * from './useNotificationListController';
 export * from './useNotifications';
 export * from './useNotificationTarget';
 export * from './useSystemNotifications';
